moved cesium viewer to molecules
