+++ web/src/components/molecules/Asset/Asset/AssetBody/Asset.tsx
@@ -2,7 +2,6 @@ import styled from "@emotion/styled";
 import { createWorldTerrain, Viewer } from "cesium";
 import { useState } from "react";
 
-import CesiumViewer from "@reearth-cms/components/atoms/CesiumViewer";
 import DownloadButton from "@reearth-cms/components/atoms/DownloadButton";
 import { DefaultOptionType } from "@reearth-cms/components/atoms/Select";
 import UserAvatar from "@reearth-cms/components/atoms/UserAvatar";
@@ -17,6 +16,7 @@ import SideBarCard from "@reearth-cms/components/molecules/Asset/Asset/AssetBody
 import UnzipFileList from "@reearth-cms/components/molecules/Asset/Asset/AssetBody/UnzipFileList";
 import ViewerNotSupported from "@reearth-cms/components/molecules/Asset/Asset/AssetBody/viewerNotSupported";
 import ArchiveExtractionStatus from "@reearth-cms/components/molecules/Asset/AssetListTable/ArchiveExtractionStatus";
+import CesiumViewer from "@reearth-cms/components/molecules/Asset/CesiumViewer";
 import { useT } from "@reearth-cms/i18n";
 import { getExtension } from "@reearth-cms/utils/file";
 import { dateTimeFormat } from "@reearth-cms/utils/format";
