+++ web/src/components/molecules/Asset/CesiumViewer/index.tsx
